Flostilux
A website for the online sale of fresh bouquets of flowers. 

The site is divided into user and administrative parts for tracking orders, budget and feedback. User-friendly, bright, pleasant to the eye and convenient in use. Any user can place an order after registration and in the future have access to their orders and their status. The user can buy
products, put them on the wish list, leave reviews, send messages and much more. 

The site is made using HTML, JavaScript, PHP and phpMyAdmin.